+++ src/ZMPRefTrajectoryGeneration/ZMPConstrainedQPFastFormulation.cpp
@@ -1237,7 +1237,7 @@ int ZMPConstrainedQPFastFormulation::BuildZMPTrajectoryFromFootTrajectory(
 
     // Compute CPU consumption time.
     gettimeofday(&end, 0);
-    CurrentCPUTime = end.tv_sec - start.tv_sec +
+    CurrentCPUTime = (double)(end.tv_sec - start.tv_sec) +
                      0.000001 * (double)(end.tv_usec - start.tv_usec);
     TotalAmountOfCPUTime += CurrentCPUTime;
     ODEBUG("Current Time : "
